Move .CAL auto loader to AMDIS plugin.
