+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri/OSGI-INF/org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.system.ChromatogramConverterAdditionService.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" configuration-policy="optional" name="org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.system.ChromatogramConverterAdditionService">
+   <service>
+      <provide interface="org.eclipse.chemclipse.converter.services.IChromatogramConverterAdditionService"/>
+   </service>
+   <implementation class="org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.service.ChromatogramConverterAdditionService"/>
+</scr:component>
