+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.xxd.calculator/META-INF/MANIFEST.MF
@@ -18,7 +18,6 @@ Bundle-ActivationPolicy: lazy
 Export-Package: org.eclipse.chemclipse.chromatogram.xxd.calculator.core.chromatogram,
  org.eclipse.chemclipse.chromatogram.xxd.calculator.core.noise,
  org.eclipse.chemclipse.chromatogram.xxd.calculator.exceptions,
- org.eclipse.chemclipse.chromatogram.xxd.calculator.io,
  org.eclipse.chemclipse.chromatogram.xxd.calculator.preferences,
  org.eclipse.chemclipse.chromatogram.xxd.calculator.settings
 Import-Package: com.fasterxml.jackson.annotation;version="2.9.2",
