+++ chemclipse/plugins/org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.chemclipse.support;bundle-version="0.8.0",
  org.eclipse.chemclipse.processing;bundle-version="0.8.0",
  org.eclipse.chemclipse.model;bundle-version="0.8.0",
- com.fasterxml.jackson.core.jackson-annotations;bundle-version="2.6.2",
  org.eclipse.chemclipse.chromatogram.xxd.calculator;bundle-version="0.8.0",
  org.eclipse.chemclipse.csd.model;bundle-version="0.9.0",
  org.eclipse.chemclipse.chromatogram.csd.identifier;bundle-version="0.9.0",
@@ -25,5 +24,6 @@ Export-Package: org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file
  org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file.model,
  org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file.preferences,
  org.eclipse.chemclipse.chromatogram.xxd.identifier.supplier.file.settings
+Import-Package: com.fasterxml.jackson.annotation;version="2.21.0"
 Bundle-Vendor: ChemClipse
 Bundle-ActivationPolicy: lazy
