+++ chemclipse/plugins/org.eclipse.chemclipse.converter/META-INF/MANIFEST.MF
@@ -13,6 +13,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.core.resources;bundle-version="3.11.1"
 Bundle-RequiredExecutionEnvironment: JavaSE-21
 Bundle-ActivationPolicy: lazy
+Bundle-Activator: org.eclipse.chemclipse.converter.Activator
 Export-Package: org.eclipse.chemclipse.converter,
  org.eclipse.chemclipse.converter.chromatogram,
  org.eclipse.chemclipse.converter.comparators,
@@ -32,6 +33,7 @@ Export-Package: org.eclipse.chemclipse.converter,
  org.eclipse.chemclipse.converter.report,
  org.eclipse.chemclipse.converter.scan,
  org.eclipse.chemclipse.converter.sequence,
+ org.eclipse.chemclipse.converter.services,
  org.eclipse.chemclipse.converter.support
 Import-Package: com.fasterxml.jackson.annotation;version="2.9.2",
  jakarta.xml.bind,
