+++ chemclipse/plugins/org.eclipse.chemclipse.msd.converter/src/org/eclipse/chemclipse/msd/converter/chromatogram/ChromatogramConverterMSD.java
@@ -14,29 +14,11 @@
  *******************************************************************************/
 package org.eclipse.chemclipse.msd.converter.chromatogram;
 
-import java.io.File;
-import java.util.Map;
-
-import org.eclipse.chemclipse.chromatogram.xxd.calculator.io.AMDISConverter;
-import org.eclipse.chemclipse.chromatogram.xxd.calculator.io.MassLibConverter;
 import org.eclipse.chemclipse.converter.chromatogram.AbstractChromatogramConverter;
 import org.eclipse.chemclipse.converter.chromatogram.IChromatogramConverter;
-import org.eclipse.chemclipse.model.columns.ISeparationColumnIndices;
-import org.eclipse.chemclipse.model.core.IScan;
-import org.eclipse.chemclipse.model.identifier.ComparisonResult;
-import org.eclipse.chemclipse.model.identifier.IComparisonResult;
-import org.eclipse.chemclipse.model.identifier.IIdentificationTarget;
-import org.eclipse.chemclipse.model.identifier.ILibraryInformation;
-import org.eclipse.chemclipse.model.identifier.LibraryInformation;
-import org.eclipse.chemclipse.model.implementation.IdentificationTarget;
-import org.eclipse.chemclipse.model.support.LibraryInformationSupport;
-import org.eclipse.chemclipse.msd.converter.preferences.PreferenceSupplier;
 import org.eclipse.chemclipse.msd.model.core.IChromatogramMSD;
 import org.eclipse.chemclipse.msd.model.core.IChromatogramPeakMSD;
-import org.eclipse.chemclipse.msd.model.core.IScanMSD;
 import org.eclipse.chemclipse.processing.DataCategory;
-import org.eclipse.chemclipse.processing.core.IProcessingInfo;
-import org.eclipse.core.runtime.IProgressMonitor;
 
 public final class ChromatogramConverterMSD extends AbstractChromatogramConverter<IChromatogramPeakMSD, IChromatogramMSD> {
 
@@ -54,114 +36,5 @@ public final class ChromatogramConverterMSD extends AbstractChromatogramConverte
 		}
 
 		return instance;
-	}
-
-	@Override
-	public void postProcessChromatogram(IProcessingInfo<IChromatogramMSD> processingInfo, IProgressMonitor monitor) {
-
-		if(processingInfo != null && processingInfo.getProcessingResult() instanceof IChromatogramMSD chromatogramMSD) {
-			parseAdditionalData(chromatogramMSD);
-		}
-	}
-
-	private void parseAdditionalData(IChromatogramMSD chromatogramMSD) {
-
-		if(chromatogramMSD != null) {
-			File file = chromatogramMSD.getFile();
-			File directory = getDirectory(file);
-			if(directory != null && directory.exists()) {
-				parseCalibrationMassLib(chromatogramMSD, directory);
-				parseTargetMassLib(chromatogramMSD, directory);
-				parseCalibrationAMDIS(chromatogramMSD, directory);
-			}
-		}
-	}
-
-	private File getDirectory(File file) {
-
-		File directory = null;
-		if(file != null) {
-			if(file.isFile()) {
-				directory = file.getParentFile();
-			} else {
-				directory = file;
-			}
-		}
-
-		return directory;
-	}
-
-	private File getFile(File directory, String fileName, String fileExtension) {
-
-		if(directory.exists()) {
-			for(File file : directory.listFiles()) {
-				if(file.isFile()) {
-					String name = file.getName().toLowerCase();
-					if(name.endsWith(fileExtension.toLowerCase()) && name.startsWith(fileName.toLowerCase())) {
-						return file;
-					}
-				}
-			}
-		}
-
-		return null;
-	}
-
-	private void parseCalibrationMassLib(IChromatogramMSD chromatogramMSD, File directory) {
-
-		if(PreferenceSupplier.isParseRetentionIndexDataMassLib()) {
-			String fileName = PreferenceSupplier.isUseChromatogramNameMassLib() ? chromatogramMSD.getName() : PreferenceSupplier.getDefaultNameMassLib();
-			File file = getFile(directory, fileName, ".inf");
-
-			if(file != null) {
-				MassLibConverter massLibConverter = new MassLibConverter();
-				IProcessingInfo<ISeparationColumnIndices> processingInfo = massLibConverter.parseRetentionIndices(file);
-				ISeparationColumnIndices separationColumnIndices = processingInfo.getProcessingResult();
-				chromatogramMSD.getSeparationColumnIndices().putAll(separationColumnIndices);
-			}
-		}
-	}
-
-	private void parseTargetMassLib(IChromatogramMSD chromatogramMSD, File directory) {
-
-		if(PreferenceSupplier.isParseTargetDataMassLib()) {
-			String fileName = PreferenceSupplier.isUseChromatogramNameMassLib() ? chromatogramMSD.getName() : PreferenceSupplier.getDefaultNameMassLib();
-			File file = getFile(directory, fileName, ".inf");
-
-			if(file != null) {
-				MassLibConverter massLibConverter = new MassLibConverter();
-				LibraryInformationSupport libraryInformationSupport = new LibraryInformationSupport();
-				String referenceIdentifierMarker = PreferenceSupplier.getReferenceIdentifierMarker();
-				String referenceIdentifierPrefix = PreferenceSupplier.getReferenceIdentifierPrefix();
-
-				IProcessingInfo<Map<Integer, String>> processingInfo = massLibConverter.parseTargets(file);
-				Map<Integer, String> targets = processingInfo.getProcessingResult();
-				for(Map.Entry<Integer, String> target : targets.entrySet()) {
-					IScan scan = chromatogramMSD.getScan(target.getKey());
-					if(scan != null && scan instanceof IScanMSD scanMSD) {
-						ILibraryInformation libraryInformation = new LibraryInformation();
-						libraryInformationSupport.extractNameAndReferenceIdentifier(target.getValue(), libraryInformation, referenceIdentifierMarker, referenceIdentifierPrefix);
-						IComparisonResult comparisonResult = ComparisonResult.COMPARISON_RESULT_BEST_MATCH;
-						IIdentificationTarget scanTargetMSD = new IdentificationTarget(libraryInformation, comparisonResult);
-						scanMSD.getTargets().add(scanTargetMSD);
-					}
-				}
-			}
-		}
-	}
-
-	private void parseCalibrationAMDIS(IChromatogramMSD chromatogramMSD, File directory) {
-
-		if(PreferenceSupplier.isParseRetentionIndexDataAMDIS()) {
-			String fileName = PreferenceSupplier.isUseChromatogramNameAMDIS() ? chromatogramMSD.getName() : PreferenceSupplier.getDefaultNameAMDIS();
-			File file = getFile(directory, fileName, ".cal");
-
-			if(file != null) {
-				AMDISConverter amdisConverter = new AMDISConverter();
-				IProcessingInfo<ISeparationColumnIndices> processingInfo = amdisConverter.parseRetentionIndices(file);
-				ISeparationColumnIndices separationColumnIndices = processingInfo.getProcessingResult();
-				chromatogramMSD.getSeparationColumnIndices().putAll(separationColumnIndices);
-			}
-		}
 	}
 }
