+++ chemclipse/tests/org.eclipse.chemclipse.chromatogram.xxd.calculator.fragment.test/META-INF/MANIFEST.MF
@@ -7,5 +7,4 @@ Bundle-Version: 0.9.0.qualifier
 Bundle-Vendor: ChemClipse
 Fragment-Host: org.eclipse.chemclipse.chromatogram.xxd.calculator;bundle-version="0.8.0"
 Import-Package: org.junit.jupiter.api;version="5.13.4"
-Require-Bundle: org.eclipse.chemclipse.converter;bundle-version="0.9.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-21
