--- chemclipse/tests/org.eclipse.chemclipse.chromatogram.xxd.calculator.fragment.test/testData/files/import/calibration1.inf
@@ -1,39 +1,0 @@
-ML.InfSavedVersion: V9.4-393
-*ANALYSIS: calibration
-*RAWFILE: calibration.D
-*DATE: 20 Mar 17  11:18 am
-*TITLE: KW 6-34
-*SAMPLE: KW 6-34
-*INSTRUMENT: MSD
-*INFO1: WAX+
-*RETIND: 2:29 800
-*RETIND: 2:50 900
-*RETIND: 3:24 1000
-*RETIND: 4:16 1100
-*RETIND: 5:24 1200
-*RETIND: 6:42 1300
-*RETIND: 8:05 1400
-*RETIND: 9:28 1500
-*RETIND: 10:49 1600
-*RETIND: 12:08 1700
-*RETIND: 13:24 1800
-*RETIND: 14:35 1900
-*RETIND: 15:45 2000
-*RETIND: 16:39 2100
-*RETIND: 17:55 2200
-*RETIND: 18:55 2300
-*RETIND: 19:59 2400
-*RETIND: 21:12 2500
-*RETIND: 22:39 2600
-*RETIND: 24:26 2700
-*RETIND: 26:41 2800
-*RETIND: 29:31 2900
-*RETIND: 33:08 3000
-*RETIND: 37:48 3100
-*RETIND: 43:51 3200
-*RETBEG: 2:30
-*RETEND: 1:08:53
-*RETBEGSCAN: 1
-*RETENDSCAN: 11087
-*RETPOL: 2
-*RETLIN: 1
