+++ chemclipse/tests/org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.fragment.test/src/org/eclipse/chemclipse/chromatogram/xxd/calculator/supplier/amdiscalri/io/AMDISConverter_1_ITest.java
@@ -11,7 +11,7 @@
  * Philip Wenig - initial API and implementation
  * Alexander Kerner - Generics
  *******************************************************************************/
-package org.eclipse.chemclipse.chromatogram.xxd.calculator.io;
+package org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.io;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -19,7 +19,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import java.io.File;
 import java.io.IOException;
 
-import org.eclipse.chemclipse.chromatogram.xxd.calculator.TestPathHelper;
 import org.eclipse.chemclipse.model.columns.IRetentionIndexEntry;
 import org.eclipse.chemclipse.model.columns.ISeparationColumnIndices;
 import org.eclipse.chemclipse.processing.core.IProcessingInfo;
@@ -29,14 +28,14 @@ import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
 
 @TestInstance(Lifecycle.PER_CLASS)
-public class AMDISConverter_2_ITest {
+public class AMDISConverter_1_ITest {
 
 	private ISeparationColumnIndices separationColumnIndices;
 
 	@BeforeAll
 	public void setUp() throws IOException {
 
-		File file = new File(TestPathHelper.TESTFILE_IMPORT_CALIBRATION_CAL_2);
+		File file = new File(TestPathHelper.TESTFILE_IMPORT_CALIBRATION_CAL_1);
 		AMDISConverter converter = new AMDISConverter();
 		IProcessingInfo<ISeparationColumnIndices> processingInfo = converter.parseRetentionIndices(file);
 		separationColumnIndices = processingInfo.getProcessingResult();
@@ -60,7 +59,7 @@ public class AMDISConverter_2_ITest {
 		IRetentionIndexEntry entry = separationColumnIndices.firstEntry().getValue();
 		assertEquals(51943, entry.getRetentionTime());
 		assertEquals(800.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 800", entry.getName());
+		assertEquals("KW800", entry.getName());
 	}
 
 	@Test
@@ -68,7 +67,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(65239);
 		assertEquals(900.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 900", entry.getName());
+		assertEquals("KW900", entry.getName());
 	}
 
 	@Test
@@ -76,7 +75,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(82126);
 		assertEquals(1000.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1000", entry.getName());
+		assertEquals("KW1000", entry.getName());
 	}
 
 	@Test
@@ -84,7 +83,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(110872);
 		assertEquals(1100.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1100", entry.getName());
+		assertEquals("KW1100", entry.getName());
 	}
 
 	@Test
@@ -92,7 +91,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(155428);
 		assertEquals(1200.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1200", entry.getName());
+		assertEquals("KW1200", entry.getName());
 	}
 
 	@Test
@@ -100,7 +99,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(214714);
 		assertEquals(1300.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1300", entry.getName());
+		assertEquals("KW1300", entry.getName());
 	}
 
 	@Test
@@ -108,7 +107,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(285142);
 		assertEquals(1400.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1400", entry.getName());
+		assertEquals("KW1400", entry.getName());
 	}
 
 	@Test
@@ -116,7 +115,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(358443);
 		assertEquals(1500.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1500", entry.getName());
+		assertEquals("KW1500", entry.getName());
 	}
 
 	@Test
@@ -124,14 +123,15 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(433899);
 		assertEquals(1600.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1600", entry.getName());
+		assertEquals("KW1600", entry.getName());
 	}
 
+	@Test
 	public void test12() {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(507559);
 		assertEquals(1700.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1700", entry.getName());
+		assertEquals("KW1700", entry.getName());
 	}
 
 	@Test
@@ -139,7 +139,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(580501);
 		assertEquals(1800.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1800", entry.getName());
+		assertEquals("KW1800", entry.getName());
 	}
 
 	@Test
@@ -147,7 +147,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(648054);
 		assertEquals(1900.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 1900", entry.getName());
+		assertEquals("KW1900", entry.getName());
 	}
 
 	@Test
@@ -155,7 +155,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(715608);
 		assertEquals(2000.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2000", entry.getName());
+		assertEquals("KW2000", entry.getName());
 	}
 
 	@Test
@@ -163,7 +163,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(777768);
 		assertEquals(2100.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2100", entry.getName());
+		assertEquals("KW2100", entry.getName());
 	}
 
 	@Test
@@ -171,7 +171,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(838854);
 		assertEquals(2200.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2200", entry.getName());
+		assertEquals("KW2200", entry.getName());
 	}
 
 	@Test
@@ -179,7 +179,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(897420);
 		assertEquals(2300.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2300", entry.getName());
+		assertEquals("KW2300", entry.getName());
 	}
 
 	@Test
@@ -187,7 +187,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(953832);
 		assertEquals(2400.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2400", entry.getName());
+		assertEquals("KW2400", entry.getName());
 	}
 
 	@Test
@@ -195,7 +195,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1008090);
 		assertEquals(2500.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2500", entry.getName());
+		assertEquals("KW2500", entry.getName());
 	}
 
 	@Test
@@ -203,7 +203,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1060908);
 		assertEquals(2600.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2600", entry.getName());
+		assertEquals("KW2600", entry.getName());
 	}
 
 	@Test
@@ -211,7 +211,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1111572);
 		assertEquals(2700.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2700", entry.getName());
+		assertEquals("KW2700", entry.getName());
 	}
 
 	@Test
@@ -219,7 +219,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1161876);
 		assertEquals(2800.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2800", entry.getName());
+		assertEquals("KW2800", entry.getName());
 	}
 
 	@Test
@@ -227,7 +227,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1221528);
 		assertEquals(2900.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 2900", entry.getName());
+		assertEquals("KW2900", entry.getName());
 	}
 
 	@Test
@@ -235,7 +235,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1294470);
 		assertEquals(3000.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 3000", entry.getName());
+		assertEquals("KW3000", entry.getName());
 	}
 
 	@Test
@@ -243,7 +243,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1386096);
 		assertEquals(3100.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 3100", entry.getName());
+		assertEquals("KW3100", entry.getName());
 	}
 
 	@Test
@@ -251,7 +251,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1502154);
 		assertEquals(3200.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 3200", entry.getName());
+		assertEquals("KW3200", entry.getName());
 	}
 
 	@Test
@@ -259,7 +259,7 @@ public class AMDISConverter_2_ITest {
 
 		IRetentionIndexEntry entry = separationColumnIndices.get(1649472);
 		assertEquals(3300.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 3300", entry.getName());
+		assertEquals("KW3300", entry.getName());
 	}
 
 	@Test
@@ -268,6 +268,6 @@ public class AMDISConverter_2_ITest {
 		IRetentionIndexEntry entry = separationColumnIndices.lastEntry().getValue();
 		assertEquals(1837398, entry.getRetentionTime());
 		assertEquals(3400.0f, entry.getRetentionIndex(), 0);
-		assertEquals("KW 3400", entry.getName());
+		assertEquals("KW3400", entry.getName());
 	}
 }
