+++ chemclipse/tests/org.eclipse.chemclipse.chromatogram.xxd.calculator.supplier.amdiscalri.fragment.test/testData/files/import/calibration1.cal
